fix bug second
