--- src/assets/data/data-btn.js
@@ -1,219 +1,0 @@
-// const data = [
-//   {
-//     id: 1,
-//     name: "iPhone 14 Pro",
-//     type: [
-//       {
-//         typeName: "128GB",
-//         price: 27999000,
-//         pricePerMonth: 1140000,
-
-//       },
-//       {
-//         typeName: "256GB",
-//         price: 30849000",
-//         pricePerMonth: 1256000,
-//       },
-//       {
-//         typeName: "512GB",
-//         price: 36549000,
-//         pricePerMonth: 1488000,
-//       },
-//       {
-//         typeName: "1TB",
-//         price: "42.249.000",
-//         pricePerMonth: "1.720.000",
-//       },
-//     ],
-//     color: {
-//       imgColor: 'purple'
-//     }
-//   },
-//   {
-//     id: 2,
-//     name: "iPhone 14 Pro Max",
-//     type: [
-//       {
-//         typeName: "128GB",
-//         price: "30.999.000",
-//         pricePerMonth: "1.262.000",
-//       },
-//       {
-//         typeName: "256GB",
-//         price: "33.849.000",
-//         pricePerMonth: "1.610.000",
-//       },
-//       {
-//         typeName: "512GB",
-//         price: "39.549.000",
-//         pricePerMonth: "1.842.000",
-//       },
-//       {
-//         typeName: "1TB",
-//         price: "45.249.000",
-//         pricePerMonth: "1.378.000",
-//       },
-//     ],
-//   },
-// ];
-
-import imgData14ProPurple from "../data-img/Slider";
-const imgPurple1 = require("../images/slider/purple/img1-purple.jpg");
-console.log("🚀 ~ file: data-btn.js:62 ~ imgPurple1:", imgPurple1);
-
-// const data1 = [
-//   {
-//     model: "iPhone14 Pro",
-//     originalCost: 27999000,
-//     description: {
-//       screen: 6.1,
-//     },
-//     option: {
-//       color: [
-//         {
-//           type: "purple",
-//           additionalFee: 0,
-//           previewImg: [
-//             purplePreviewImg1,
-//             purplePreviewImg2,
-//             purplePreviewImg13,
-//           ],
-//           media: {
-//             pos_1: purplePos1Img,
-//             pos_2: purplePos2Img,
-//             pos_3: purplePos3Img,
-//           },
-//         },
-//         {
-//           type: "gold",
-//           additionalFee: 0,
-//           previewImg: [goldPreviewImg1, goldPreviewImg2, goldPreviewImg13],
-//           media: {
-//             pos_1: goldPos1Img,
-//             pos_2: goldPos2Img,
-//             pos_3: goldPos3Img,
-//           },
-//         },
-//         {
-//           type: "silver",
-//           additionalFee: 0,
-//           previewImg: [
-//             silverPreviewImg1,
-//             silverPreviewImg2,
-//             silverPreviewImg13,
-//           ],
-//           media: {
-//             pos_1: silverPos1Img,
-//             pos_2: silverPos2Img,
-//             pos_3: silverPos3Img,
-//           },
-//         },
-//         {
-//           type: "black",
-//           additionalFee: 0,
-//           previewImg: [blackPreviewImg1, blackPreviewImg2, blackPreviewImg13],
-//           media: {
-//             pos_1: blackPos1Img,
-//             pos_2: blackPos2Img,
-//             pos_3: blackPos3Img,
-//           },
-//         },
-//       ],
-//       storage: [
-//         {
-//           capacity: "128GB",
-//           additionalFee: 0,
-//         },
-//         {
-//           capacity: "256GB",
-//           additionalFee: 200,
-//         },
-//         {
-//           capacity: "512GB",
-//           additionalFee: 200,
-//         },
-//         {
-//           capacity: "1TB",
-//           additionalFee: 200,
-//         },
-//       ],
-//     },
-//   },
-//   {
-//     model: "iPhone14 Pro Max",
-//     originalCost: 30999000,
-//     description: {
-//       screen: 6.7,
-//     },
-//     option: {
-//       color: [
-//         {
-//           type: "purple",
-//           additionalFee: 0,
-//           previewImg: [
-//             purplePreviewImg1,
-//             purplePreviewImg2,
-//             purplePreviewImg13,
-//           ],
-//           media: {
-//             pos_1: purplePos1Img,
-//             pos_2: purplePos2Img,
-//             pos_3: purplePos3Img,
-//           },
-//         },
-//         {
-//           type: "gold",
-//           additionalFee: 0,
-//           previewImg: [goldPreviewImg1, goldPreviewImg2, goldPreviewImg13],
-//           media: {
-//             pos_1: goldPos1Img,
-//             pos_2: goldPos2Img,
-//             pos_3: goldPos3Img,
-//           },
-//         },
-//         {
-//           type: "silver",
-//           additionalFee: 0,
-//           previewImg: [
-//             silverPreviewImg1,
-//             silverPreviewImg2,
-//             silverPreviewImg13,
-//           ],
-//           media: {
-//             pos_1: silverPos1Img,
-//             pos_2: silverPos2Img,
-//             pos_3: silverPos3Img,
-//           },
-//         },
-//         {
-//           type: "black",
-//           additionalFee: 0,
-//           previewImg: [blackPreviewImg1, blackPreviewImg2, blackPreviewImg13],
-//           media: {
-//             pos_1: blackPos1Img,
-//             pos_2: blackPos2Img,
-//             pos_3: blackPos3Img,
-//           },
-//         },
-//       ],
-//       storage: [
-//         {
-//           capacity: "128GB",
-//           additionalFee: 0,
-//         },
-//         {
-//           capacity: "256GB",
-//           additionalFee: 200,
-//         },
-//         {
-//           capacity: "512GB",
-//           additionalFee: 200,
-//         },
-//         {
-//           capacity: "1TB",
-//           additionalFee: 200,
-//         },
-//       ],
-//     },
-//   },
-// ];
