+++ src/assets/data/index.js
@@ -1,31 +1,29 @@
-import { createSlice } from "@reduxjs/toolkit";
-
 import {
   purplePreviews,
   goldPreviews,
   silverPreviews,
   blackPreviews,
-} from "../../../components/ScrollContent/importAssets";
+} from "../../components/ScrollContent/importAssets";
 import {
   purplePos1Img,
   purplePos2Img,
   purplePos3Img,
-} from "../../../components/ScrollContent/importAssets";
+} from "../../components/ScrollContent/importAssets";
 import {
   goldPos1Img,
   goldPos2Img,
   goldPos3Img,
-} from "../../../components/ScrollContent/importAssets";
+} from "../../components/ScrollContent/importAssets";
 import {
   silverPos1Img,
   silverPos2Img,
   silverPos3Img,
-} from "../../../components/ScrollContent/importAssets";
+} from "../../components/ScrollContent/importAssets";
 import {
   blackPos1Img,
   blackPos2Img,
   blackPos3Img,
-} from "../../../components/ScrollContent/importAssets";
+} from "../../components/ScrollContent/importAssets";
 
 const devices = [
   {
@@ -186,43 +184,4 @@ function initData(device) {
   }
   return result;
 }
-
-var initial = initData(devices[0]);
-
-export const selectSlice = createSlice({
-  name: "counter",
-  initialState: { value: initial },
-  reducers: {
-    handlClickDevice: (state, action) => {
-      state.value = initData(action.payload);
-    },
-    selectColor: (state, action) => {
-      state.value = {
-        ...state.value,
-        option: {
-          ...state.value.option,
-          selected_color: action.payload,
-          added_colorFee: action.payload.additionalFee,
-        },
-      };
-    },
-
-    selecStorage: (state, action) => {
-      state.value = {
-        ...state.value,
-        option: {
-          ...state.value.option,
-          selected_storage: action.payload,
-        },
-      };
-    },
-  },
-});
-
-export const {
-  incrementByAmount,
-  handlClickDevice,
-  selectColor,
-  selecStorage,
-} = selectSlice.actions;
-export default selectSlice.reducer;
+export { devices, initData };
